Bug fixing when adding a watch brand.
