--- src/main/java/com/wristwatch/domain/WatchBrandSearchForm.java
@@ -1,19 +1,0 @@
-package com.wristwatch.domain;
-
-/**
- * Created by web on 19/04/17.
- */
-public class WatchBrandSearchForm {
-
-    String brandname;
-
-    public String getBrandname() {
-        return brandname;
-    }
-
-    public void setBrandname(String firstname) {
-        this.brandname = brandname;
-    }
-
-
-}
