+++ src/main/java/com/wristwatch/domain/WatchbrandRepository.java
@@ -1,0 +1,10 @@
+package com.wristwatch.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+/**
+ * Created by web on 24/04/17.
+ */
+public interface WatchbrandRepository extends JpaRepository<Watchbrand, Long> {
+
+}
