+++ src/main/resources/templates/fragments/header.html
@@ -17,7 +17,7 @@
     </div>
     <div class="nav">
     <ul>
-        <li><a th:href="@{/watchbrand/index}">Brands</a></li>
+        <li><a th:href="@{/watchbrand/all}">Brands</a></li>
         <li><a th:href="@{/user/search}">Search</a></li>
         <li><a th:href="@{/user/dashboard}">Dashboard</a></li>
         <li><a th:href="@{/user/logout}">Logout</a></li>
