--- src/main/resources/templates/watchbrand/addWatchBrand.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <th:block th:include="fragments/header :: headerinc"></th:block>
-</head>
-<body>
-
-<div th:include="fragments/header :: header"></div>
-
-<div class="content">
-<h3>Add Brand</h3>
-
-<form action ="#" th:action="@{/watchbrand/addbrand}" th:object="${watchBrand}" method="post">
-    <input type="hidden" th:field="*{id}"/>
-    <p>Brand Name: </p> <input type="text" th:field="*{brandname}"/>
-
-    <input type="submit" value="Add Brand" />
-</form>
-
-<div th:if="${message!=null}">
-    <p/>
-    <span th:text="${message}"/>
-</div>
-</div>
-
-</body>
-</html>
