+++ src/main/resources/templates/watchbrand/index.html
@@ -6,40 +6,23 @@
 <body>
 
 <div th:include="fragments/header :: header"></div>
-
 <div class="content">
-<!--<table>-->
-    <!--<tr>-->
-        <!--<th>Brand Name</th>-->
-        <!--<th>Update</th>-->
-        <!--<th>Delete</th>-->
-    <!--</tr>-->
-    <!--<tr th:each="watchBrand : ${watchBrands}">-->
-        <!--<td th:text="${watchBrand.brandname}">Fin</td>-->
-        <!--<td><a href="#" th:href="@{/watchbrand/update/}+${watchBrand.id}">Update</a></td>-->
-        <!--<td><a href="#" th:href="@{/watchbrand/delete/}+${watchBrand.id}">Delete</a></td>-->
-    <!--</tr>-->
-<!--</table>-->
-
-
-<div>
-<ol>
-    <li th:each="watchBrand : ${watchBrands}">
-        <span th:text="${watchBrand.firstname}+ ' '"></span>
-        <a href="#" th:href="@{/watchbrand/update/}+${watchBrand.id}">Update</a>
-        <a href="#" th:href="@{/watchbrand/delete/}+${watchBrand.id}">Delete</a>
-    </li>
-</ol>
+    <div>
+        <ol>
+            <li th:each="watchbrand : ${watchbrands}">
+                <span th:text="${watchbrand.brandname}">Rolex</span>
+                <a href="#" th:href="@{watchbrand/delete/}+${watchbrand.id}">Delete</a>
+            </li>
+        </ol>
+    </div>
 </div>
 
-    <div class="form">
-    <form action ="#" th:action="@{/watchbrand/addbrand}" th:object="${watchBrand}" method="post">
+<div class="form">
+    <form action="#" th:action="@{/watchbrand/addbrand}" th:object="${watchbrand}" method="get">
+
         <button type="submit">Add Brand</button>
     </form>
-    </div>
-
 </div>
-
 
 </body>
 </html>
