+++ src/main/resources/templates/watchbrand/updateWatchBrand.html
@@ -9,7 +9,7 @@
 
 <h3>Watch Brand Update</h3>
 
-<form action ="#" th:action="@{/watchbrand/update}" th:object="${watchBrand}" method="post">
+<form action ="#" th:action="@{/watchbrand/update}" th:object="${watchBrand1}" method="post">
     <input type="hidden" th:field="*{id}"/>
     <p>First Name: </p> <input type="text" th:field="*{brandname}"/>
